+++ packages/rdgis/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(rdgis)
+
+test_check("rdgis")